Add authentication layer to the Content API
In order to request content, clients now need to use a OAuth2
Bearer token with scope 'read'.
The token has to be created manually by someone with admin access
as the authorization flow has not been publicly exposed.

The migration files create a default Frontend Application and
Access Token to speed things up, the token is randomly generated.