Make all live and draft links open in a new win
Some live and view draft links were opening in a new window and
some weren't for no apparent reason. This fixes that inconsistency.

I opened a PR to the Wagtail main repo
https://github.com/torchbox/wagtail/pull/3211 as there were some
other html/python files to change and it was too hacky to include
it in our repo.

We should delete the `pages_for_moderation` and `recent_edits` pages
if/when that PR is merged and a new version of Wagtail released.

This also changes the way edit previews work.
Before each revision would open a new tab and this was confusing
people as the number of tabs kept increasing.
The preview window now is unique per page, meaning that changing the
page multiple time and clicking on the preview button now open
the preview page in the same window.